% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_fv.R
\name{aggregate_fv}
\alias{aggregate_fv}
\title{Aggregate \link[spatstat.explore]{fv.object}s by Cluster}
\usage{
aggregate_fv(
  X,
  by = stop("must specify `by`"),
  f_aggr_ = c("mean", "median", "max", "min"),
  ...
)
}
\arguments{
\item{X}{a \link[spatstat.geom]{hyperframe},
containing one or more \link[spatstat.explore]{fv.object} column(s)}

\item{by}{one-sided \link[stats]{formula}, sample clustering.
Use only one-level hierarchy (e.g., \code{~patient} or \code{~image}).
Do not use multi-level hierarchy (e.g., \code{~patient/image})}

\item{f_aggr_}{\link[base]{character} scalar, method to aggregate
within cluster, currently supports
\code{'mean'}, \code{'median'}, \code{'max'}, and \code{'min'}.}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{aggregate_fv} returns a \link[base]{data.frame}, with
aggregated information stored in \link[base]{matrix}-columns.

Note that \link[spatstat.geom]{hyperframe} does not support
\link[base]{matrix}-column (for good reasons!).
Therefore, function \link{aggregate_fv} must return \link[base]{data.frame},
instead of \link[spatstat.geom]{hyperframe}.
}
\description{
Aggregate information in \link[spatstat.explore]{fv.object}s
by sample clustering.
}
