% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_num.R
\name{aggregate_num}
\alias{aggregate_num}
\alias{aggregate_quantile}
\alias{aggregate_kerndens}
\title{Aggregate \link[base]{numeric} \link[spatstat.geom:hyperframe]{hypercolumns} and/or \link[spatstat.geom]{marks}, by Cluster}
\usage{
aggregate_num(
  X,
  by = stop("must specify `by`"),
  FUN,
  f_aggr_ = c("mean", "median", "max", "min"),
  ...
)

aggregate_quantile(X, ...)

aggregate_kerndens(X, ...)
}
\arguments{
\item{X}{a \link[spatstat.geom]{hyperframe},
containing either or all of
\itemize{
\item {one or more \link[base]{numeric}
\link[spatstat.geom:hyperframe]{hypercolumns}}
\item {one-and-only-one \link[spatstat.geom]{ppp}-\link[spatstat.geom:hyperframe]{hypercolumns}
with one or more \link[base]{numeric} \link[spatstat.geom]{marks}}
}}

\item{by}{one-sided \link[stats]{formula},
one-level hierarchy clustering, e.g., \code{~patient} or \code{~image}.
Do \strong{not} use multi-level hierarchy, e.g., \code{~patient/image}}

\item{FUN}{\link[base]{function} to extract information, currently supports
functions \link[stats]{quantile} and \link{kerndens}}

\item{f_aggr_}{\link[base]{character} scalar, method to aggregate
within cluster, currently supports
\code{'mean'}, \code{'median'}, \code{'max'}, and \code{'min'}.}

\item{...}{additional parameters of function \code{FUN}}
}
\value{
Function \link{aggregate_num} returns a \link[base]{data.frame}, with
aggregated information stored in \link[base]{matrix}-columns.
}
\description{
Aggregate \link[base]{numeric} \link[spatstat.geom:hyperframe]{hypercolumns}
and/or \link[spatstat.geom]{marks}
by sample clustering.
}
\details{
Function \link{aggregate_quantile} is a wrapper of
workhorse function \link{aggregate_num} with \code{FUN = quantile}.

Function \link{aggregate_kerndens} is a wrapper of
workhorse function \link{aggregate_num} with \code{FUN = kerndens}.
}
