% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvlist.R
\name{check_fvlist}
\alias{check_fvlist}
\title{check_fvlist}
\usage{
check_fvlist(X, ...)
}
\arguments{
\item{X}{a \link[base]{list} of \link[spatstat.explore]{fv.object}s}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{check_fvlist} does not have a returned value.
}
\description{
check_fvlist
}
\details{
To check
\itemize{
\item {if \eqn{x}-axis of all \link[spatstat.explore]{fv.object}s are all same}
\item {\code{attr(,'fname')} of all \link[spatstat.explore]{fv.object}s are all same}
}

Note that
\itemize{
\item {function \link{key1.fv} returns of all \link[spatstat.explore]{fv.object}s are not required to be all same}
}
}
\keyword{internal}
