% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvlist.R
\name{fvlist}
\alias{fvlist}
\alias{key1val.fvlist}
\alias{cumtrapz.fvlist}
\title{Operations on a \link[base]{list} of \link[spatstat.explore]{fv.object}s}
\usage{
key1val.fvlist(X, check = TRUE, ...)

cumtrapz.fvlist(X, check = TRUE, ...)
}
\arguments{
\item{X}{a \link[base]{list} of \link[spatstat.explore]{fv.object}s}

\item{check}{\link[base]{logical} scalar, an option to suppress
function \link{check_fvlist} in a batch process.
Default \code{TRUE}}

\item{...}{additional parameters, currently not in use}
}
\value{
All functions return a \link[base]{double} \link[base]{matrix}.
}
\description{
Operations on a \link[base]{list} of \link[spatstat.explore]{fv.object}s
}
\details{
Function \link{key1val.fvlist} gathers the primary outcome (via \link{key1.fv})
of the \link[spatstat.explore]{fv.object}s.

Function \link{cumtrapz.fvlist} is a batch process of function \link{cumtrapz.fv}.
}
