% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_hyperframe.R
\name{getGroupsFormula.grouped_hyperframe}
\alias{getGroupsFormula.grouped_hyperframe}
\title{Extract Grouping Formula from \code{grouped_hyperframe}}
\usage{
\method{getGroupsFormula}{grouped_hyperframe}(object, asList, sep)
}
\arguments{
\item{object}{a \code{grouped_hyperframe}}

\item{asList, sep}{place holders for S3 generic \link[nlme]{getGroupsFormula}}
}
\value{
Function \link{getGroupsFormula.grouped_hyperframe} returns a one-sided \link[stats]{formula}
}
\description{
..
}
\keyword{internal}
