% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupedHyperframe.R
\name{getGroupsFormula.groupedHyperframe}
\alias{getGroupsFormula.groupedHyperframe}
\title{Extract Grouping Formula from \code{groupedHyperframe}}
\usage{
\method{getGroupsFormula}{groupedHyperframe}(object, asList, sep)
}
\arguments{
\item{object}{a \code{groupedHyperframe}}

\item{asList, sep}{place holders for S3 generic \link[nlme]{getGroupsFormula}}
}
\value{
Function \link{getGroupsFormula.groupedHyperframe} returns a one-sided \link[stats]{formula}
}
\description{
..
}
\keyword{internal}
