% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupedHyperframe.R
\name{groupedHyperframe}
\alias{groupedHyperframe}
\title{Grouped Hyperframe}
\value{
A \link{groupedHyperframe}
}
\description{
A class \link{groupedHyperframe} to represent \link[spatstat.geom]{hyperframe}
with a (multilevel) hierarchical structure.
}
\details{
The class \link{groupedHyperframe} inherits from class \link[spatstat.geom]{hyperframe}.
This class has additional \link[base]{attributes}
\describe{
\item{\code{attr(,'group')}}{\link[stats]{formula}}
}
}
