% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fv_S3.R
\name{is.finite.fv}
\alias{is.finite.fv}
\title{Finite Indices of \link[spatstat.explore]{fv.object}}
\usage{
\method{is.finite}{fv}(x)
}
\arguments{
\item{x}{an \link[spatstat.explore]{fv.object}}
}
\value{
Function \link{is.finite.fv} returns a \link[base]{logical} \link[base]{vector}.
}
\description{
Finite indices of the \emph{black solid curve}
as shown in \link[spatstat.explore]{plot.fv}
of an \link[spatstat.explore]{fv.object}.
}
\details{
Function \link{is.finite.fv} is an S3 method dispatch of the generic function
\link[base]{is.finite}.

Function \link{is.finite.fv} finds the \link[base]{finite},
i.e., non-\code{NA}, non-\code{NaN} and non-\code{Inf},
indices of the \emph{black solid curve} of an \link[spatstat.explore]{fv.object}.
}
