% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key1nonfinite.fv.R
\name{key1nonfinite<-}
\alias{key1nonfinite<-}
\title{Assign a Value to \code{NaN} and/or \code{Inf} in \link[spatstat.explore]{fv.object}}
\usage{
key1nonfinite(x) <- value
}
\arguments{
\item{x}{an \link[spatstat.explore]{fv.object}}

\item{value}{\link[base]{numeric} scalar, to replace
the \code{NaN} and \code{Inf} elements. Default \code{0}.}
}
\description{
..
}
\examples{
library(spatstat.data)
library(spatstat.explore)
# ?markcorr
r1 = seq(from = 0, to = 25, by = 1)
r2 = seq(from = 0, to = 250, by = 10)
markcorr(spruces, r = r1) |> key1val.fv()
markcorr(spruces, r = r2) |> key1val.fv()

\dontrun{
debug(sewsmod); markcorr(spruces, r = r2)} # reason found

a = markcorr(spruces, r = r2)
key1val.fv(a)
key1nonfinite(a) = 0
key1val.fv(a)

# equivalent
markcorr(spruces, r = r2) |>
 `key1nonfinite<-`(value = 0) |>
 key1val.fv()
}
