% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_S3.R
\name{log.ppp}
\alias{log.ppp}
\title{log.ppp}
\usage{
\method{log}{ppp}(x, base = exp(1))
}
\arguments{
\item{x}{a \link[spatstat.geom]{ppp.object}}

\item{base}{\link[base]{numeric} scalar}
}
\value{
Function \link{log.ppp} returns a \link[spatstat.geom]{ppp.object}.
}
\description{
...
}
\details{
Function \link{log.ppp} takes a \link[base]{log} of continuous marks
of a \link[spatstat.geom]{ppp.object}.
}
\examples{
data(longleaf, package = 'spatstat.data')
plot(longleaf)
plot(log.ppp(longleaf))
}
