% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_.R
\name{nested_}
\alias{nested_}
\title{Nested \link[base]{factor}s}
\usage{
nested_(lang, data, sep = ".", lex.order = TRUE)
}
\arguments{
\item{lang}{any \link[base]{language} of a nested structure, e.g., \code{~c1/c2} or \code{c1/c2}}

\item{data}{\link[base]{data.frame}}

\item{sep}{see function \link[base]{interaction}}

\item{lex.order}{see function \link[base]{interaction}, default \code{TRUE}}
}
\value{
Function \link{nested_} returns a \link[base]{factor} (from function \link[base]{interaction}) with additional \link[base]{attributes},
\describe{
\item{\code{attr(.,'group1')}}{\link[stats]{formula}, the highest grouping}
}
}
\description{
..
}
\details{
Function \link{nested_} ..
}
\note{
R base function \link[base]{interaction} correctly handles syntactically invalid names (see \link[base]{make.names}).
}
\examples{
data(farms, package = 'MASS')
interaction(farms[c('Mois', 'Manag')])

(f = nested_(~ Mois/Manag, data = farms))
stopifnot(identical(f, nested_(quote(~ Mois/Manag), data = farms)))
stopifnot(identical(f, nested_(quote(Mois/Manag), data = farms)))
}
\keyword{internal}
