% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nncross_.R
\name{nncross_}
\alias{nncross_}
\alias{nncross_.ppp}
\title{Alternative Interface of \link[spatstat.geom]{nncross}}
\usage{
nncross_(X, ...)

\method{nncross_}{ppp}(X, i, j, ...)
}
\arguments{
\item{X}{see \strong{Details}}

\item{...}{additional parameters,
including \code{i} and \code{j} for function \link{nncross_.hyperframe}}

\item{i, j}{\link[base]{character} or \link[base]{integer} scalars.
See functions \link[spatstat.explore]{Gcross},
and \link[spatstat.explore]{Kcross} for more details}
}
\value{
Function \link{nncross_.ppp} returns
a \link[base]{numeric} \link[base]{vector}
if \code{i} and \code{j} are valid mark values of \link[spatstat.geom]{ppp.object} \code{X};
otherwise returns a \code{NULL} value.
}
\description{
An alternative interface of function \link[spatstat.geom]{nncross}.
}
\details{
Function \link{nncross_.ppp} creates an interface similar to
functions \link[spatstat.explore]{Gcross},
\link[spatstat.explore]{Jcross},
and \link[spatstat.explore]{Kcross}, etc.,
which takes an \link[spatstat.geom]{is.multitype} \link[spatstat.geom]{ppp.object}
and two mark values \code{i} and \code{j},
then calls the workhorse function
\link[spatstat.geom]{nncross} with parameter \code{what = 'dist'}.
If mark values \code{i} and \code{j} does not exist in the \link[spatstat.geom]{ppp.object},
a \code{NULL} value will be returned.
}
\examples{
library(spatstat.data)
library(spatstat.geom)

(xs = split.ppp(amacrine))
(a1 = nncross(X = xs$off, Y = xs$on, what = 'dist'))
a2 = nncross_.ppp(amacrine, i = 'off', j = 'on')
a3 = nncross_.ppp(amacrine, i = 1L, j = 2L)
stopifnot(identical(a1, a2), identical(a1, a3))

nncross_.ppp(amacrine, i = 'a', j = 'b') # exception handling
}
