% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_ppplist.R
\name{op_ppplist}
\alias{op_ppplist}
\alias{fv_ppplist}
\alias{dist_ppplist}
\title{Batch Operations of \code{'ppplist'} Object}
\usage{
fv_ppplist(x, ...)

dist_ppplist(x, ...)
}
\arguments{
\item{x}{a \code{'ppplist'} object}

\item{...}{additional parameters of workhorse functions
\link{fv_ppp} or \link{dist_ppp}}
}
\value{
Function \link{fv_ppplist} returns a \link[base]{list} of function \link{fv_ppp} returns.

Function \link{dist_ppplist} returns a \link[base]{list} of function \link{dist_ppp} returns.
}
\description{
Batch operations of function \link{fv_ppp} or \link{dist_ppp}, for a \code{'ppplist'} input.
}
\details{
Function \link{fv_ppplist} is a batch process of the workhorse function \link{fv_ppp},
with the help of R-core package \pkg{parallel}.

Function \link{dist_ppplist} is a batch process of the workhorse function \link{dist_ppp},
with the help of R-core package \pkg{parallel}.
}
\examples{
library(spatstat.data)
library(spatstat.geom) # for ?spatstat.geom::split.ppp
library(spatstat.explore) # for ?spatstat.explore::Emark, etc.

# create a 'ppplist' object
Vc = with(shapley$marks, expr = {
 cut.default(V, breaks = quantile(V, probs = c(0, 1/3, 2/3, 1)), labels = c('L', 'M', 'H'))
})
x0 = shapley; x0$marks$V = NULL
plot(x <- split.ppp(x0, f = Vc))

xE = fv_ppplist(x, fn = Emark)
names(xE)
names(xE$L)
stopifnot(inherits(xE$L$Mag.E, what = 'fv'))

# create another 'ppplist' object
nbf = nbfires |> 
  subset.ppp(select = c('fire.type', 'cause', 'ign.src')) |>
  na.omit.ppp()
x2 = split.ppp(nbf, f = nbf$marks$fire.type)
dist_ppplist(x2, fn = nncross_.ppp, i = 'rrds', j = 'ltning')
dist_ppplist(x2, fn = nncross_.ppp, i = 'unknown', j = 'burn.no.perm')
}
\keyword{internal}
