% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fv_S3.R
\name{op_fv}
\alias{op_fv}
\alias{key1.fv}
\alias{trapz.fv}
\alias{cumtrapz.fv}
\title{Operations on \link[spatstat.explore]{fv.object}}
\usage{
key1.fv(X, ...)

trapz.fv(X, ...)

cumtrapz.fv(X, ...)
}
\arguments{
\item{X}{an \link[spatstat.explore]{fv.object}}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{key1.fv} returns a \link[base]{character} scalar.

Function \link{trapz.fv} returns a \link[base]{numeric} scalar.

Function \link{cumtrapz.fv} returns a \link[base]{numeric} \link[base]{vector}.
}
\description{
Various operations on a function value \link[spatstat.explore]{fv.object}.
}
\details{
Function \link{key1.fv} finds the name of primary outcome,
i.e., the \emph{black solid curve} as shown in \link[spatstat.explore]{plot.fv},
of an \link[spatstat.explore]{fv.object}.

Functions \link{trapz.fv} and \link{cumtrapz.fv}
obtain the (cumulative)
\link[pracma]{trapz}oidal integration of
the area under the \emph{\strong{black solid curve}} (see function \link{key1.fv})
of a function value \link[spatstat.explore]{fv.object}.
}
\examples{
library(spatstat.data)
library(spatstat.explore)
library(spatstat.geom) # ?spatstat.geom::subset.ppp

# numeric mark
plot(x <- Emark(spruces))
key1.fv(x)
trapz.fv(x)
tail(cumtrapz.fv(x))

plot(x <- Vmark(spruces))
trapz.fv(x)
tail(cumtrapz.fv(x))

plot(x <- markcorr(spruces))
trapz.fv(x)
tail(cumtrapz.fv(x))

plot(x <- markvario(spruces))
trapz.fv(x)
tail(cumtrapz.fv(x))

# multitype
(btc = subset.ppp(betacells, select = 'type'))
plot(x <- Gcross(btc, i = 'off', j = 'on'))
key1.fv(x)
trapz.fv(x)
tail(cumtrapz.fv(x))

plot(x <- Kcross(btc, i = 'off', j = 'on'))
trapz.fv(x)
tail(cumtrapz.fv(x))

plot(x <- Lcross(btc, i = 'off', j = 'on'))
trapz.fv(x)
tail(cumtrapz.fv(x))

plot(x <- Jcross(btc, i = 'off', j = 'on'))
x$km # um.. bug?
#trapz.fv(x)
#tail(cumtrapz.fv(x))

plot(x <- pcfcross(btc, i = 'off', j = 'on'))
head(x$iso)
head(is.finite(x))
trapz.fv(x)
tail(cumtrapz.fv(x))

plot(x <- markconnect(btc, i = 'off', j = 'on'))
trapz.fv(x)
tail(cumtrapz.fv(x))

# roc
plot(x <- roc.ppp(swedishpines, covariate = 'x'))
trapz.fv(x)
tail(cumtrapz.fv(x))
}
\keyword{internal}
